Temporary patch to fix Null Pointers In merging castings

+++ grakn-graph/src/main/java/ai/grakn/graph/internal/AbstractGraknGraph.java
@@ -866,6 +866,9 @@ public abstract class AbstractGraknGraph<G extends Graph> implements GraknGraph,
             //Fix the duplicates
             Set<Relation> duplicateRelations = mergeCastings(mainCasting, duplicated);
 
+            //Temporary patch to stop null pointers. No proper fix due to castings being removed in master
+            if(duplicateRelations.isEmpty()) return false;
+
             //Remove Redundant Relations
             duplicateRelations.forEach(relation -> ((ConceptImpl) relation).deleteNode());
 
@@ -886,6 +889,10 @@ public abstract class AbstractGraknGraph<G extends Graph> implements GraknGraph,
      * @return A set of possible duplicate relations.
      */
     private Set<Relation> mergeCastings(CastingImpl mainCasting, Set<CastingImpl> castings){
+
+        //Temporary patch to stop null pointers. No proper fix due to castings being removed in master
+        if(mainCasting == null) return Collections.emptySet();
+
         RoleType role = mainCasting.getRole();
         Set<Relation> relations = mainCasting.getRelations();
         Set<Relation> relationsToClean = new HashSet<>();
